Revisited and Updated the LPQ experiment
The code is now generic in a sense, that it accepts any number of models
and plotting. The legend is missing though. The results on the Yale
Facedatabase B show, that I can't reproduce the results in the paper,
with the same experimental setup. This is quite disappointing, but
finally I was able to close this missing experiment.