New Feature added, Webservice extended
A new feature has been added: Resize! What a simple operator, but so
useful. This means, you can learn a new model with the Resize operator
as a feature and you never need to rescale incoming images by yourself.

The Webservice has been extended with a new PredictableModelWrapper and
a NumericalDataSet,. The numerical dataset enables you to use strings in
a PredictableModel for your predictions, something I wanted to have for
a long time.

These methods will be added to the framework as soon as possible.